chore: improve error messages when type assertions fail in tests
Prior to this commit, if one of our type assertions failed in
a test, we didn't get any information about the cause of the error.
This commit changes the assertions to use the signature that allows
us to provide an error message, and in the error message we include
the return value so that it is visible for debugging.