feat: improvements to how collection update ttls are specified (#45)
This commit introduces a new class, CollectionTtl, to represent
the different use cases for dealing with TTLs when updating
collection objects.  It encapsulates the `refreshTtl` and `ttl`
arguments into a single structure and provides an API for
constructing and modifying them which will hopefully make
it more clear to users what the semantics are.