refactor: use hit/miss/error pattern for dictremovefields response (#63)
We refactor CacheDictionaryGetFieldsResponse to use the Hit/Miss/Error pattern. In doing so we remove the need for the previous payload accessors. When a dictionary is not found, we return CacheDictionaryGetFieldsResponse.Miss instead of Success with unary response objects that are all Miss.

We also add a payload accessor convenience similar to the one in CacheDictionaryFetchResponse. The end-user can retrieve the payload as a dictionary. Note in particular: when a queried field is missing from the cache, we do not include it in the dictionary. An alternative would be to include the field in the dictionary and associate to it a null value, which is allowed in .NET. Since we cannot store null values in the cache, we instead adopt the convention cache miss implies not present.