chore: refactor ScsDataClient for consistency of `Send` functions
This commit contains no functional changes, just re-organizes
the code in ScsDataClient to make it simpler to maintain in
the future.  Changes include:

* We had a lot of helper functions that were intended for internal
  use; some were prefixed with the name `Send` and some were not.
  These are now all prefixed with the name `Send`.
* Some of the `Send` methods were private and some were not.  They
  are now all private.
* Some of the `Send` methods accepted optional arguments, which was
  not necessary.  They no longer accept optional arguments.
* Some of the public methods delegated to `Send` methods, others did
  not.  They now all delegate to `Send` methods.
* The private `Send` methods are all now moved to the bottom of the
  file so that the code organization is more consistent.

This will make it easier to maintain this code in the future, especially
when we need to apply mechanical changes.  For example, in a subsequent
PR I will be adding logging statements to all of the `Send` methods.
This will be much easier and less error-prone now that all of the methods
that require it are in one place in the code.